+++ benchmarks/golden_configs/oss_mnist.py
@@ -7,7 +7,7 @@
 def get_golden_real_stats():
 
     return {
-        "reference_speed": 660,
+        "reference_speed": 578,
         "reference_memory": 945,
         "reference_loss": 0.026,
     }